Flow collection phase
