+++ crates/typst/src/model/quote.rs
@@ -6,7 +6,7 @@ use crate::foundations::{
 };
 use crate::introspection::Locatable;
 use crate::layout::{
-    Alignment, BlockChild, BlockElem, Em, HElem, PadElem, Spacing, VElem,
+    Alignment, BlockBody, BlockElem, Em, HElem, PadElem, Spacing, VElem,
 };
 use crate::model::{CitationForm, CiteElem};
 use crate::text::{SmartQuoteElem, SmartQuotes, SpaceElem, TextElem};
@@ -185,7 +185,7 @@ impl Show for Packed<QuoteElem> {
 
         if block {
             realized = BlockElem::new()
-                .with_body(Some(BlockChild::Content(realized)))
+                .with_body(Some(BlockBody::Content(realized)))
                 .pack()
                 .spanned(self.span());
